+++ src/main/java/domain/card/Type.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Type {
     DIAMOND("다이아몬드"),
