+++ src/main/java/domain/gambler/Names.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.gambler;
 
 import java.util.Arrays;
 import java.util.List;
