+++ src/test/java/domain/card/CardDeckTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.card;
 
+import domain.card.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
